pyproject.toml
[tool.poetry]
name = "siim-isic-melanoma-classification"
version = "0.1.0"
description = ""
authors = ["Gianluca Rossi <gr.gianlucarossi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.0.4"
scikit_learn = "^0.23.1"
mlxtend = "^0.17.2"
lightgbm = "^2.3.1"
torch = "^1.5.0"
torchvision = "^0.6.0"
matplotlib = "^3.2.1"
pillow-simd = "^7.0.0"
pytorch-lightning = "^0.7.6"
opencv-python = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
kaggle = "^1.5.6"
dvc = "^0.94.0"
black = "^19.10b0"
mypy = "^0.770"
flake8 = "^3.8.2"
ipdb = "^0.13.2"
ipython = "^7.14.0"
rope = "^0.17.0"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
